renderer: Make `Renderer_Uniform_Buffer`'s actually work
Right now Uniform Buffers are created with legacy usage API, where it
can only be suggested what type of operation and its frequency would be
on a buffer.
At the same time, the mapping interface (accessing the
underlying buffer stored on GPU) uses modern access API, which relies
on usage of access flags/bits.  As well as standard Read/Write's, it
exposes flags such as `PERSISTENT`, `COHERENT`, and others.
Currently, it is expected that Uniform Buffers are handled with legacy
API logic until the proper modern API buffer creation and management is
implemented in the engine.

Also, old 'usage' enum fields are replaced with ones directly mapped to
OpenGL values.  Use cases other than of OpenGL will be handled later,
for now it is easier to debug and see values this way.