entity_table: New module
Introduces `Entity_Lookup_Table` struct.  It has two 'paired' arrays of
`Entity_ID`s and `Entity` pointers to entity storages, where the actual
entity data is stored.  It can be thought of as an Key-Value Map.
On the first iteration, the underlying storage was an `Array< Entity_Key_Value >`,
where one `Entity_Key_Value` object held both the id and the pointer.
The problem is that when searching for an entity, you only compare
against ids and the pointer is accessed once in the end, thus loading
unused field every time.  That is why the approach was switched from
AoS (Array of Structs) to SoA (Struct of Arrays).
The underlying storage can be changed in future based on performance,
use-cases, etc.