allocator: Remove `NULL` pointer `Assert`s in `reallocate` & `deallocate`
This is so the `stb_image` library works.
The asserts are removed because it is well defined that:
 - In case of calling C function `realloc` with `NULL` pointer, it is
expected to be the same as calling `malloc`.  This does happen in the
`stb_image` library.
 - In case of calling C function `free` with `NULL` pointer, it is expected
to do nothing.  It is probably a good idea to not even call it, who knows
how much work it does until it returns.