app: in 'Entities' window, make `Entity`'s model ID changeable
It was supposed to be a nested view into a `Model` instance, but that
seems to be too complicated to include in a single window.  Make it so
it is a single scalar input of 'Model ID' that can be changed - the
appropriate model will be resolved right away since the model instance
is resolved from model ID every draw command.