G-Buffer: Invert values for Roughness maps used as Specular maps
A small tweak that allows Roughness textures - a part of PBR material
workflow - to be used in Blinn-Phong lighting model - a non-PBR workflow.
Both Roughness and Specularity refer to the same property of a material,
but starts from the opposite "ends" of the scale.  Basically, a roughness
of 1 is equal to a specularity of 0 - and so on.

Be careful!
A regular specular map texture will be inverted too! Again, this is just
a temporary solution so PBR Roughness textures could be used as-is.